+++ src/features/ProductImageMagnify/RenderEnlargedImage.tsx
