+++ src/features/ProductImageMagnify/hint/DefaultHint.tsx
