+++ src/features/ProductImageMagnify/hint/DisplayUntilActive.tsx
