+++ src/features/ProductImageMagnify/lens/negative-space/Lens.tsx
