+++ src/features/ProductImageMagnify/lens/negative-space/LensBottom.tsx
