+++ src/features/ProductImageMagnify/lens/negative-space/LensLeft.tsx
