+++ src/features/ProductImageMagnify/lens/negative-space/LensRight.tsx
