+++ src/features/ProductImageMagnify/lens/negative-space/LensTop.tsx
