+++ src/features/ProductImageMagnify/lens/negative-space/index.tsx
