+++ src/features/ProductImageMagnify/lens/positive-space/assets/texture.gif
