+++ src/features/ProductImageMagnify/lens/positive-space/assets/textured-lens-data-uri.ts
