+++ src/features/ProductImageMagnify/lib/dimensions.ts
