+++ src/features/ProductImageMagnify/lib/imageCoordinates.ts
