+++ src/features/ProductImageMagnify/lib/imageRatio.ts
