+++ src/features/ProductImageMagnify/lib/styles.ts
