+++ src/features/ProductImageMagnify/prop-types/EnlargedImage.ts
