+++ src/features/ProductImageMagnify/prop-types/Image.ts
