+++ src/features/ProductImageMagnify/prop-types/Lens.ts
