+++ src/features/ProductImageMagnify/prop-types/Point.ts
