+++ src/features/ProductImageMagnify/prop-types/index.ts
