+++ src/features/ProductImageMagnify/utils/detectIt.ts
